Metrics

Types

SequencerJ currently supports the four algorithms originally cited in the article and supported by the python version of the Sequencer. These are:

Distances.SqEuclidean — Type

SqEuclidean([thresh])

Create a squared-euclidean semi-metric. For the meaning of thresh, see Euclidean.

Missing docstring.

Missing docstring for KLDivergence. Check Documenter's build log for details.

SequencerJ.EMD — Type

EMD

Earth Mover's Distance, a.k.a. 1-p Monge-Wasserstein distance. The algorithm as implemented here assumes balanced transport where both vectors have the same statistical mass (they both sum to 1). Input vectors should be normalized to sum to 1.

SequencerJ.Energy — Type

Energy

Energy distance as defined by Székely. An explicit grid may be provided. Default is axes(u,1) and axes(v,1).

Energy distance

Each algorithm has associated to it a constant that represents a usable default type instance (with a default grid):

Constants

SequencerJ.L2 — Constant

Constant to use when specifying the Square Euclidean or L2 distance metric for a sequencing run.

SequencerJ.KLD — Constant

Constant for specifiying the Kullbach-Leibler Divergence metric.

SequencerJ.WASS1D — Constant

Constant for specifiying the Monge-Wasserstein a.k.a. 1-p Wasserstein a.k.a. Earth Mover's Distance (EMD) metric. This metric is sensitive to the underlying grid. Default is unit grid taken from the axes of the data vector.

SequencerJ.ENERGY — Constant

Constant for specifiying the Energy distance as defined by Székely. Wikipedia article

SequencerJ.ALL_METRICS — Constant

Convenience constant to represent all of L2, WASS1D, KLD, and ENERGY.

Usage

Data Orientation

SequencerJ operates on pairwise columns of data in the given matrix (or vector of vectors). It builds an N x N distance matrix for further analysis, given an N-column input. It is therefore important to properly orient the data matrix so that M observations run downward along N columns, with each column representing a separate data set or series.

This can be performed efficiently with permutedims:

A = rand(10,5)
M,N = size(A)
A = M < N ? permutedims(A) : A

10×5 Array{Float64,2}:
 0.99952    0.668349   0.457608  0.0567948  0.203346
 0.232961   0.681261   0.976968  0.9045     0.448153
 0.598168   0.640356   0.178708  0.615234   0.811661
 0.698737   0.0115577  0.282144  0.722056   0.986782
 0.162458   0.677213   0.619702  0.728158   0.156785
 0.690132   0.152384   0.581529  0.519223   0.565625
 0.663307   0.0212311  0.410208  0.0455111  0.65151
 0.026314   0.513785   0.103104  0.0580772  0.512276
 0.292633   0.232851   0.948449  0.84673    0.335721
 0.0980011  0.406992   0.998342  0.500926   0.316422

While a 2-D data set such as a photograph may be shuffled and reassembled in either orientation, the Sequencer algorithm will work better with more rows and fewer columns (i.e. M > N).

Autoscaling

The Sequencer algorithm's processing time rises as N x N, posing a challenge for data exploration with larger data sets. It may require considerable experimentation with multiple scales and metrics before finding an optimum combination. To assist with finding an appropriate scale, SequencerJ offers an "autoscaling" option. This is, in fact, the default behavior with the keyword argument scales=nothing.

In autoscaling, SequencerJ first performs the sequencing function at several scales on a subset of columns. The scale producing the greatest elongation (see elong) is then used on the full dataset. This is a feature specific to the Julia port of the Sequencer algorithm.

At present, there is no "autometric" option to choose the best metric.

Metrics and Grids

Setting the metrics keyword option as

sequence(A, metrics=(L2, ENERGY))

means to use Distances.SqEuclidean() and SequencerJ.Energy().

With equal effect, you can write:

sequence(A, metrics=(SqEuclidean(), Energy())

You can specify a threshold for the SqEuclidean calculations. To use, e.g., 1e-7 means round to zero any distance calculation results smaller than 1e-7.

sequence(A, metrics=(SqEuclidean(1e-7), Energy())

To specify a non-unit data grid, set grid as a Vector{AbstractFloat} in the range [0.0-1.0) and use the Distance type, not the constant.

using SequencerJ #hide
A = rand(100,100);
M,N = size(A)
g = collect(0.0:0.5:N-1)
sequence(A, metrics=(Energy()), grid=g)

To these four metrics, the intention is to add spectral methods for complex (i.e. 2-D) inputs and any other Distances that make sense for sequencing 1-D or 2-D data. I would love community help with this!